Fix tz1 key verification
Move address hashing into ssi so it can be accessed in ldp.rs.
Cannot use to_did_method for this.
