+++ Cargo.toml
@@ -40,6 +40,8 @@ http = { version = "0.2", optional = true }
 serde_urlencoded = "0.7"
 percent-encoding = { version = "2.1", optional = true }
 tokio = { version = "1.0", optional = true, features = ["macros"] }
+blake2b_simd = "0.5"
+bs58 = { version = "0.4", features = ["check"] }
 
 [workspace]
 members = [
