+++ did-tezos/Cargo.toml
@@ -13,5 +13,3 @@ tokio = { version = "1.0", features = ["macros", "rt"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 async-trait = "0.1"
-blake2b_simd = "0.5"
-bs58 = { version = "0.4", features = ["check"] }
