+++ src/ldp.rs
@@ -2,14 +2,15 @@ use async_trait::async_trait;
 use chrono::prelude::*;
 
 // use crate::did::{VerificationMethod, VerificationMethodMap};
-use crate::did::{Resource, Source};
+use crate::blakesig;
+use crate::did::Resource;
 use crate::did_resolve::{dereference, Content, DIDResolver, DereferencingInputMetadata};
 use crate::error::Error;
 use crate::hash::sha256;
 use crate::jwk::{Algorithm, OctetParams as JWKOctetParams, Params as JWKParams, JWK};
 use crate::rdf::DataSet;
 use crate::urdna2015;
-use crate::vc::{Context, Contexts, LinkedDataProofOptions, Proof};
+use crate::vc::{LinkedDataProofOptions, Proof};
 use serde_json::Value;
 
 // TODO: factor out proof types
@@ -289,10 +290,9 @@ impl ProofSuite for Ed25519BLAKE2BDigestSize20Base58CheckEncodedSignature2021 {
     async fn verify(
         proof: &Proof,
         document: &(dyn LinkedDataDocument + Sync),
-        resolver: &dyn DIDResolver,
+        _resolver: &dyn DIDResolver,
     ) -> Result<(), Error> {
         let jws = proof.jws.as_ref().ok_or(Error::MissingProofSignature)?;
-        let didtz = resolver.to_did_method().ok_or(Error::MissingKey)?;
         let jwk: JWK = match proof.property_set {
             Some(ref props) => {
                 let jwk_value = props.get("publicKeyJwk").ok_or(Error::MissingKey)?;
@@ -300,10 +300,15 @@ impl ProofSuite for Ed25519BLAKE2BDigestSize20Base58CheckEncodedSignature2021 {
             }
             None => return Err(Error::MissingKey),
         };
-        let did = didtz
-            .generate(&Source::Key(&jwk))
-            .ok_or(Error::MissingKey)?;
-        if Some(did + "#blockchainAccountId") != proof.verification_method {
+        // Ensure the verificationMethod corresponds to the hashed public key.
+        // The hash should be in the vM URL.
+        let vm = proof
+            .verification_method
+            .as_ref()
+            .ok_or(Error::MissingVerificationMethod)?;
+        let hash = blakesig::hash_public_key(&jwk)?;
+        let vm_base = vm.split('#').next().unwrap();
+        if vm_base.split(':').find(|&h| h == &hash).is_none() {
             return Err(Error::KeyMismatch);
         }
         let message = to_jws_payload(document, proof).await?;
